Revert "refactor: unify Core Audio property-read boilerplate into shared helpers"
